api/requirements.txt
fastapi
numpy
requests
Pillow
uvicorn
dwave-system